// src/types/stats.ts
export interface UserStats {
  id: string;
  ip: string;
  location: string;
  language: string;
  browser: string;
  operating_system: string;
  time_zone: string;
  screen_resolution: string;
  user_agent: string;
  form_data: any;
  generated_letter: string;
  generation_type: string;
  referral_code: string;
  created_at: string;
}